# contracts/openapi/mercatus-v1.yaml
openapi: 3.1.0
info:
  title: AIBank — riel de pago para agentes
  version: "1.0.0"
  summary: Autorizar y capturar pagos entre sub-cuentas de una plataforma de agentes.
  description: |
    Implementación del contrato `aibank_integration_contract.md` v0.3 de Mercatus,
    con los defaults acordados en `docs/10-preguntas-mercatus.md` y confirmados en
    la respuesta de Mercatus.

    ## Unidad monetaria

    **Todo importe es un entero en millonésimas (10⁻⁶) de la unidad mayor.**

    | Importe | Valor |
    |---|---|
    | $0.001 | `1000` |
    | $0.50 | `500000` |
    | $1.00 | `1000000` |

    Seis decimales es el máximo representable. Un importe con más precisión se
    rechaza en vez de truncarse en silencio.

    ## Dos modelos de integración

    | | **A — ómnibus** | **B — cuenta propia** |
    |---|---|---|
    | Titular de la cuenta pagadora | La plataforma | El cliente final del banco |
    | Quién autoriza | La plataforma, con su credencial | El titular, con un mandato |
    | Cómo se indica | Sin `mandate_id` | Con `mandate_id` |

    **El contrato de pago es idéntico en los dos.** Lo único que cambia es de
    dónde sale la autoridad para mover el dinero, y se expresa con un campo. Ver
    `POST /v1/consent-requests` para el flujo de permiso del modelo B.

    ## Autorizar y después capturar

    `POST /v1/authorizations` reserva el dinero en la cuenta pagadora sin
    entregarlo. `POST /v1/authorizations/{id}/capture` lo mueve de verdad. Entre
...

...

...

...

...

...
servers:
...
security:
  - bearerToken: []
...
paths:
...
  /v1/authorizations/{authorizationId}:
    get:
      tags: [Pagos]
      summary: Consulta el estado de una autorización
      operationId: getAuthorization
      description: |
        La llama el **vendedor**, no el comprador: es la verificación independiente
        de que el pago existe y coincide antes de entregar lo pagado.

        Los parámetros `expected_*` son opcionales. Si se envían y no coinciden, la
        respuesta es `422` en vez del estado — así el vendedor no tiene que
        comparar de su lado, aunque puede seguir haciéndolo si prefiere.
      security:
        - bearerToken: [payments:read]
      parameters:
        - name: authorizationId
          in: path
          required: true
          schema: { type: string }
          example: auth_9c2f1e8843a24b5c9d0e1f2a3b4c5d6e
        - name: expected_amount
          in: query
          schema: { type: integer, format: int64 }
          description: Monto que el vendedor espera, en millonésimas.
        - name: expected_payee_account_id
          in: query
          schema: { type: string }
          description: Cuenta receptora que el vendedor espera.
      responses:
        "200":
          description: Estado de la autorización.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthorizationDetail" }
              example:
                authorization_id: auth_9c2f1e8843a24b5c9d0e1f2a3b4c5d6e
                payer_account_id: acc_7f3e9a2b4c1d4e5f8a9b0c1d2e3f4a5b
                payee_account_id: acc_1c9d44f0a1b2c3d4e5f60718293a4b5c
                amount: 1000
                currency: USD
                status: captured
                expires_at: "2026-09-01T18:57:00Z"
                settled_at: "2026-09-01T18:42:12Z"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/AuthorizationNotFound" }
        "422":
          description: |
            `amount_mismatch` o `recipient_mismatch`: lo autorizado no coincide con
            lo que el vendedor esperaba.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /v1/mandates/{mandateId}:
    get:
      tags: [Consentimiento]
      summary: Consulta un permiso y cuánto queda de él
      operationId: getMandate
      description: |
        Conviene consultarlo: saber cuánto queda permite pedir una ampliación
        **antes** de que un pago falle delante de un usuario.

        `consumed` se deriva de las autorizaciones vivas, no de un contador. Una
        retención sin capturar ya consume el tope; liberarla o reembolsarla lo
        devuelve.
      security:
        - bearerToken: [payments:read]
      parameters:
        - name: mandateId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: El permiso.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Mandate" }
              example:
                mandate_id: 4a7b1c2d-9e8f-4a1b-8c2d-3e4f5a6b7c8d
                account_id: acc_3f9e8d7c6b5a4938271605f4e3d2c1b0
                currency: USD
                status: active
                max_per_operation: 5000000
                max_total: 20000000
                consumed: 8000000
                remaining: 12000000
                expires_at: "2026-10-01T20:57:00Z"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: |
            `mandate_not_found`. Misma respuesta si no existe o si fue otorgado a
            otra integración.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/RateLimited" }
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Motivo legible por máquina. Estable; compáralo por igualdad.
        message:
          type: string
          description: Explicación para un humano. NO es estable; no lo parsees.
...
    AuthorizationStatus:
      type: string
      enum:
        - authorized
        - captured
        - partially_refunded
        - refunded
        - released
        - expired
      description: |
        - `authorized`: dinero retenido, todavía no entregado.
        - `captured`: liquidado al receptor.
        - `refunded` / `partially_refunded`: devuelto al pagador.
        - `released`: la retención se devolvió sin cobrar.
        - `expired`: la vigencia terminó. No se puede capturar.
    Authorization:
      type: object
      required: [authorization_id, status]
      properties:
        authorization_id:
          type: string
          pattern: "^auth_[0-9a-f]{32}$"
        status: { $ref: "#/components/schemas/AuthorizationStatus" }
        expires_at:
          type: string
          format: date-time
        settled_at:
          type: string
          format: date-time
          description: Presente solo tras capturar.
    AuthorizationDetail:
      allOf:
        - $ref: "#/components/schemas/Authorization"
        - type: object
          properties:
            payer_account_id: { type: string }
            payee_account_id: { type: string }
            amount: { type: integer, format: int64 }
            currency: { type: string }
...
    Mandate:
      type: object
      required: [mandate_id, account_id, currency, status, consumed, expires_at]
      properties:
        mandate_id: { type: string }
        account_id: { type: string }
        currency: { type: string }
        status:
          type: string
          enum: [active, revoked, expired]
        max_per_operation: { type: integer, format: int64 }
        max_total: { type: integer, format: int64 }
        consumed:
          type: integer
          format: int64
          description: Comprometido hasta ahora, incluidas las retenciones vivas.
        remaining:
          type: integer
          format: int64
          description: Ausente si el mandato no tiene tope total.
        expires_at: { type: string, format: date-time }
        revoked_at: { type: string, format: date-time }
...
  responses:
    Unauthorized:
      description: |
        `invalid_credential`: falta el token, expiró o no es válido. Pide uno nuevo.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: |
        `insufficient_scope`: el token es válido pero no incluye el scope
        necesario. Pedir otro token con el mismo `client_id` NO lo resuelve.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    AuthorizationNotFound:
      description: "`authorization_not_found`."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RateLimited:
      description: |
        `rate_limited`. Espera lo que indique `Retry-After` y reintenta con la
        MISMA `Idempotency-Key`.
      headers:
        Retry-After:
          schema: { type: integer }
          description: Segundos.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...